requirements.txt
numpy
pyserial
mediapipe
opencv-python